+++ pyirena/__init__.py
@@ -29,7 +29,7 @@ References:
     Beaucage, G. (1996). J. Appl. Cryst. 29, 134-146
 """
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 __author__ = "Jan Ilavsky"
 __email__ = "ilavsky@aps.anl.gov"
 
@@ -38,7 +38,8 @@ from pyirena.core.sizes import SizesDistribution
 from pyirena.core.simple_fits import SimpleFitModel, MODEL_REGISTRY, MODEL_NAMES
 from pyirena.batch import (
     fit_unified, fit_sizes, fit_simple, fit_pyirena,
-    manipulate_data, average_data, fit_saxs_morph, pxp_to_nexus,
+    manipulate_data, average_data, fit_saxs_morph,
+    igor_to_nexus, pxp_to_nexus,
 )
 from pyirena.io.results import load_result, SUPPORTED_ANALYSES
 
@@ -62,6 +63,7 @@ __all__ = [
     "fit_saxs_morph",
     "manipulate_data",
     "average_data",
+    "igor_to_nexus",
     "pxp_to_nexus",
     "load_result",
     "SUPPORTED_ANALYSES",